VTKU-110 : Historian pakitus omaan luokkaan
Ei toiminnallisia muutoksia.
